Pull `CoordinatesFactory` up
